+++ linking-platform/smart-card/src/hoverCard.ts
@@ -1,0 +1,2 @@
+export { HoverCard } from './view/HoverCard';
+export type { HoverCardProps } from './view/HoverCard/types';