+++ linking-platform/smart-card/src/utils/analytics/index.ts
@@ -11,6 +11,7 @@ export {
   unresolvedEvent,
   invokeSucceededEvent,
   invokeFailedEvent,
+  chunkloadFailedEvent,
   connectSucceededEvent,
   connectFailedEvent,
   trackAppAccountAuthStarted,