+++ linking-platform/smart-card/src/utils/index.tsx
@@ -269,3 +269,24 @@ export const handleOnClick =
 
 export const isIntersectionObserverSupported = () =>
   typeof IntersectionObserver !== 'undefined';
+
+export const sleep = (ms: number): Promise<void> => {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+};
+
+export const importWithRetry = async <T,>(
+  importFn: () => Promise<T>,
+  retries = 2,
+  interval = 500,
+): Promise<T> => {
+  try {
+    return await importFn();
+  } catch (error) {
+    if ((error as Error).name === 'ChunkLoadError' && retries > 0) {
+      await sleep(interval);
+      return importWithRetry(importFn, retries - 1, interval);
+    } else {
+      throw error;
+    }
+  }
+};